Fix warnings

+++ Open-BMBR/model_animation.h
@@ -175,7 +175,7 @@ private:
 		auto& boneInfoMap = m_BoneInfoMap;
 		int& boneCount = m_BoneCounter;
 
-		for (int boneIndex = 0; boneIndex < mesh->mNumBones; ++boneIndex)
+		for (unsigned int boneIndex = 0; boneIndex < mesh->mNumBones; ++boneIndex)
 		{
 			int boneID = -1;
 			std::string boneName = mesh->mBones[boneIndex]->mName.C_Str();
@@ -198,7 +198,7 @@ private:
 
 			for (int weightIndex = 0; weightIndex < numWeights; ++weightIndex)
 			{
-				int vertexId = weights[weightIndex].mVertexId;
+				long unsigned int vertexId = weights[weightIndex].mVertexId;
 				float weight = weights[weightIndex].mWeight;
 				assert(vertexId <= vertices.size());
 				SetVertexBoneData(vertices[vertexId], boneID, weight);
